Split fixture/pytest tests.
